[CASCL-1323] Move CFN-stack helpers out of `apply/install_mode.go`
`InstallModeTagKey` and `DetectedInstallMode` are CloudFormation-stack
concerns, not pflag.Value plumbing. Putting them in `install_mode.go`
made that file diverge from `inference_method.go` and
`create_karpenter_resources.go`, which both moved over from `install`
with only a package rename.

Move them to `run.go` next to `KarpenterStackName` /
`DDKarpenterStackName`, where the rest of the CFN logic lives. The
three pflag-type files in `apply/` are now structurally symmetric:
data + pflag methods, nothing else.
